+++ plugins/video/loader/bmp/bmpimage.h
@@ -41,7 +41,7 @@ class csBMPImageIO : public iImageIO
   virtual ~csBMPImageIO ();
 
   virtual const csImageIOFileFormatDescriptions& GetDescription ();
-  virtual csPtr<iImage> Load (uint8* iBuffer, uint32 iSize, int iFormat);
+  virtual csPtr<iImage> Load (uint8* iBuffer, size_t iSize, int iFormat);
   virtual void SetDithering (bool iEnable);
   virtual csPtr<iDataBuffer> Save (iImage *image, const char *mime = 0,
     const char* extraoptions = 0);
@@ -66,13 +66,13 @@ class ImageBMPFile : public csImageFile
 {
   friend class csBMPImageIO;
   // Load a Windows-format BMP file
-  bool LoadWindowsBitmap (uint8* iBuffer, uint32 iSize);
+  bool LoadWindowsBitmap (uint8* iBuffer, size_t iSize);
 
 private:
   /// Initialize the image object
   ImageBMPFile (int iFormat) : csImageFile (iFormat) { };
   /// Try to read the BMP file from the buffer and return success status
-  bool Load (uint8* iBuffer, uint32 iSize);
+  bool Load (uint8* iBuffer, size_t iSize);
 };
 
 #endif // __CS_BMPIMAGE_H__
