+++ plugins/video/loader/gif/gifimage.h
@@ -40,7 +40,7 @@ class csGIFImageIO : public iImageIO
   virtual ~csGIFImageIO ();
 
   virtual const csImageIOFileFormatDescriptions& GetDescription ();
-  virtual csPtr<iImage> Load (uint8* iBuffer, uint32 iSize, int iFormat);
+  virtual csPtr<iImage> Load (uint8* iBuffer, size_t iSize, int iFormat);
   virtual void SetDithering (bool iEnable);
   virtual csPtr<iDataBuffer> Save (iImage *image, const char *mime = 0,
     const char* extraoptions = 0);
@@ -64,7 +64,7 @@ private:
   /// Initialize the image object
   ImageGifFile (int iFormat) : csImageFile (iFormat) { };
   /// Try to read the GIF file from the buffer and return success status
-  bool Load (uint8* iBuffer, uint32 iSize);
+  bool Load (uint8* iBuffer, size_t iSize);
 };
 
 #endif // __CS_GIFIMAGE_H__
