+++ plugins/video/loader/mplex/imgplex.h
@@ -57,7 +57,7 @@ class csMultiplexImageIO : public iImageIO
 
   virtual bool Initialize (iObjectRegistry*);
   virtual const csImageIOFileFormatDescriptions& GetDescription ();
-  virtual csPtr<iImage> Load (uint8* iBuffer, uint32 iSize, int iFormat);
+  virtual csPtr<iImage> Load (uint8* iBuffer, size_t iSize, int iFormat);
   virtual void SetDithering (bool iEnable);
   virtual csPtr<iDataBuffer> Save (iImage *image, const char *mime = 0,
     const char* extraoptions = 0);
