+++ plugins/video/loader/tga/tgaimage.h
@@ -40,7 +40,7 @@ class csTGAImageIO : public iImageIO
   virtual ~csTGAImageIO ();
 
   virtual const csImageIOFileFormatDescriptions& GetDescription ();
-  virtual csPtr<iImage> Load (uint8* iBuffer, uint32 iSize, int iFormat);
+  virtual csPtr<iImage> Load (uint8* iBuffer, size_t iSize, int iFormat);
   virtual void SetDithering (bool iEnable);
   virtual csPtr<iDataBuffer> Save (iImage *image, const char *mime = 0,
     const char* extraoptions = 0);
@@ -65,7 +65,7 @@ private:
   /// Initialize the image object
   ImageTgaFile (int iFormat) : csImageFile (iFormat) { };
   /// Try to read the PNG file from the buffer and return success status
-  bool Load (uint8* iBuffer, uint32 iSize);
+  bool Load (uint8* iBuffer, size_t iSize);
 };
 
 #endif // __CS_TGAIMAGE_H__
